Issue #839 Move network error helpers to validation folder
