+++ src/stores/services/actions.ts
@@ -2,7 +2,7 @@ import { Id as TopicId } from '../../fixtures/types/topics';
 import * as constants from '../../application/constants';
 import * as helpers from '../helpers/make_action';
 import { Service } from '../../validation/services/types';
-import { Errors } from '../../errors/types';
+import { Errors } from '../../validation/errors/types';
 import { LatLong } from '../../validation/search/types';
 
 export type BuildTopicServicesRequestAction = Readonly<ReturnType<typeof buildTopicServicesRequestAction>>;
